new rotation + starting on new animations and ui

+++ data/DualPieceShape.cs
@@ -17,7 +17,7 @@ public class DualPieceShape : IHasShape
         get => R.B;
     }
 
-    public DualPieceShape((int, int)[] shape, string name, int weight = 6)
+    public DualPieceShape((float, float)[] shape, string name, int weight = 6)
     {
         L = new(shape, name, 0, false);
         R = L.GetFlippedShape();
